+++ Source/Engine/Renderer/Renderer.h
@@ -19,7 +19,7 @@ namespace kiko
 		void BeginFrame();
 		void EndFrame();
 
-		void SetColor(int R, int G, int B, int A);
+		void SetColor(uint8_t R, uint8_t G, uint8_t B, uint8_t A);
 		void DrawLine(int x1, int y1, int x2, int y2);
 		void DrawLine(float x1, float y1, float x2, float y2);
 		void DrawPoint(int x, int y);
